Added classe animale and fixed some types

+++ src/main/java/it/unibo/adozione_animali/model/api/Spazio.java
@@ -2,11 +2,11 @@ package it.unibo.adozione_animali.model.api;
 
 public interface Spazio {
 
-    boolean insertSpazio(int spazio, String tipo, int dimensione);
+    boolean insertSpazio(int spazio, String tipo, float dimensione);
 
     boolean updateSpazioTipo(int spazio, String newTipo);
 
-    boolean updateSpazioDimensione(int spazio, int newDimensione);
+    boolean updateSpazioDimensione(int spazio, float newDimensione);
 
     boolean deleteSpazio(int spazio);
 }
