+++ src/main/java/it/unibo/adozione_animali/model/api/SpazioPersona.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 public interface SpazioPersona {
 
-    boolean insertSpazioPersona(int spazio, String persona, String provincia, String citta, int numero, Optional<Integer> spazioRimanente);
+    boolean insertSpazioPersona(int spazio, String persona, String provincia, String citta, int numero, Optional<Float> spazioRimanente);
 
     boolean updateSpazioRimanente(int spazio, String persona, String provincia, String citta, int numero, int newSpazio);
